Relative calls, jumps: specially handle offset of 0.
Offset of 0 means that control flow is transferred inside a (next, IIRC)
instruction, which doesn't make sense and complicates control flow recovery.
On the other hand, 0 offset is how undefined, unrelocated destinations are
represented in .o files. Consider this latter case more realistic, so
disassemble in such a way as to avoid inside-instruction jumps (specifically,
address used in disassembly is address of the instruction itself).