xtensa: Don't perform side-effect operations in ana().
Don't change data type (like, making something an offset) in ana(), as it
may lead to disassembly listing changed just by browsing it (because ana()
is called to render an instruction). Instead, do this in emu(), which is
called once during initial analysis (that means that data type won't be
automagically propagated to where it's referenced by l32r, but listing
stability is more important, and doing the above is something for plugin
anyway).